\docType{package}
\name{gpindexExtra-package}
\alias{gpindexExtra-package}
\alias{gpindexExtra}

\title{Extras for Generalized Price and Quantity Indexes}

\description{
Extra functionality to enhance the Generalized Price and Quantity Indexes package.
}

\seealso{
\url{https://github.com/marberts/gpindexExtra}
}

\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com}
}